util: rename io.py into util.py
