+++ schulze_condorcet/util.py
