+++ tests/test_schulze.py
@@ -4,7 +4,7 @@ from typing import Dict, List, Optional, Tuple, TypedDict
 import unittest
 
 from schulze_condorcet import schulze_evaluate, schulze_evaluate_detailed
-import schulze_condorcet.io as io
+import schulze_condorcet.util as util
 from schulze_condorcet.strength import margin, winning_votes
 from schulze_condorcet.types import Candidate, DetailedResultLevel as DRL, VoteString
 
@@ -518,10 +518,10 @@ class MyTest(unittest.TestCase):
         condensed = schulze_evaluate(reference_votes, candidates)
         self.assertEqual("1=0>2", condensed)
 
-    def test_io(self) -> None:
+    def test_util(self) -> None:
         candidates = ["1", "2", "3"]
         # This does only static type conversion
-        self.assertEqual(io.as_candidates(candidates), candidates)
+        self.assertEqual(util.as_candidates(candidates), candidates)
 
         # build the same votes, represented as string, list and tuple
         vote_str_1 = "1"
@@ -534,18 +534,18 @@ class MyTest(unittest.TestCase):
         vote_list_list = [vote_list_1, vote_list_2]
         vote_tuple_list = [vote_tuple_1, vote_tuple_2]
 
-        self.assertEqual(vote_str_list, io.as_vote_strings(vote_list_list))
-        self.assertEqual(vote_str_list, io.as_vote_strings(vote_tuple_list))
+        self.assertEqual(vote_str_list, util.as_vote_strings(vote_list_list))
+        self.assertEqual(vote_str_list, util.as_vote_strings(vote_tuple_list))
 
         # mixing different representations of votes is confusing and therefore not
         # recommended and forbidden by static type checking.
         # However, we ensure the outcome is right nonetheless.
         self.assertEqual(
             2*vote_str_list,
-            io.as_vote_strings([*vote_list_list, *vote_tuple_list])  # type:ignore
+            util.as_vote_strings([*vote_list_list, *vote_tuple_list])  # type:ignore
         )
 
-        self.assertEqual(vote_tuple_list, io.as_vote_tuples(vote_str_list))
+        self.assertEqual(vote_tuple_list, util.as_vote_tuples(vote_str_list))
 
 
 if __name__ == '__main__':
